requirements.txt
Django==4.2.11
django-ckeditor==6.7.1
django-filter==24.3
django-js-asset==2.2.0
django-multiselectfield==0.1.13
django-phonenumber-field==8.0.0
djangorestframework==3.15.2